Rename derectory lang to locales
