+++ code/js/indexComponents/helpSubpages/documentation.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import mainPage_lang from "../../lang/mainPage_lang";
-import {productName} from "../../lang/mainPage_lang";
+import mainPage_lang from "../../locales/mainPage_lang";
+import {productName} from "../../locales/mainPage_lang";
 
 
 // The component is responsible for displaying the documentation page
